finished 4.23

+++ osa4/controllers/blogs.js
@@ -1,47 +1,53 @@
 const blogRouter = require('express').Router()
 const Blog = require('../models/blog')
+const User = require('../models/user')
+const jwt = require('jsonwebtoken')
 
-blogRouter.get('/', (req, res) => {
-    Blog
-        .find({})
-        .then(blogs => {
-            res.json(blogs)
-        })
+
+
+blogRouter.get('/', async (req, res) => {
+    res.json(await Blog.find({}).populate('user'))
 })
 
-blogRouter.post('/', (req, res, next) => {
-    const blogPost = new Blog(req.body)
-    blogPost.save()
-        .then(result => {
-            res.status(201).json(result)
-        })
-        .catch(error => next(error))
+blogRouter.post('/', async (req, res, next) => {
+    if (!req.user) {
+        return res.status(401).json({ error: 'Token invalid' })
+    }
+    const user = await User.findById(req.user)
+    const blogPost = new Blog({
+        title: req.body.title,
+        author: req.body.author,
+        url: req.body.url,
+        likes: req.body.likes,
+        user: user._id
+    })
+    const savedBlog = await blogPost.save()
+    user.blogs = user.blogs.concat(savedBlog._id)
+    await user.save()
+    res.status(201).json(savedBlog)
 })
 
 blogRouter.delete('/:id', async (req, res, next) => {
+    const blog = await Blog.findById(req.params.id)
+    if (blog === null) {
+        return res.status(404).json({ error: 'Blog not found' })
+    }
+    if (!req.user || req.user !== blog.user.toString()) {
+        return res.status(401).json({ error: 'Token invalid' })
+    }
     await Blog.findByIdAndRemove(req.params.id)
     res.status(204).end()
 })
 
-blogRouter.put('/:id', async (req, res, next)=>{
+blogRouter.put('/:id', async (req, res, next) => {
     const blog = {
         title: req.body.title,
         author: req.body.author,
         url: req.body.url,
         likes: req.body.likes,
     }
-    const response = await Blog.findByIdAndUpdate(req.params.id, blog, {new: true})
+    const response = await Blog.findByIdAndUpdate(req.params.id, blog, { new: true })
     res.json(response)
 })
-
-const errorHandler = (error, req, res, next) => {
-    if (error.name === 'ValidationError') {
-        return res.status(400).send(error.message).end()
-    } else {
-        return res.status(502).end()
-    }
-}
-
-blogRouter.use(errorHandler)
 
 module.exports = blogRouter
